Fix various compiler warnings Unused variables Signed/unsigned mismatches Assignment within conditional expression Conditional expression is constant Cast truncates constant value Declaration hides previous local declaration Declaration hides function parameter Declaration hides class member Declaration hides global declaration Unreferenced function with internal linkage has been removed Potentially uninitialized local variable used Unreachable code

+++ src/game/client/sound/SoundCache.cpp
@@ -202,7 +202,7 @@ std::optional<std::tuple<ALint, ALint>> SoundCache::TryLoadCuePoints(const std::
 
 	data.resize(size);
 
-	if (fread(data.data(), 1, size, file.get()) != size)
+	if (fread(data.data(), 1, size, file.get()) != static_cast<std::size_t>(size))
 	{
 		m_Logger->error("Error reading file \"{}\"", fileName);
 		return {};
